% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrun.R
\name{nnmodel}
\alias{nnmodel}
\title{Assemble neural network model definitions with datasets to form a 'nnmodel object'.}
\usage{
nnmodel(model_name, scenario_name = "", climate_source = "",
  static_filename, dynamic_filename, dynamic_colnames = NA,
  dynamic_filename2 = NA, dynamic_colnames2 = NA, field_order = NA,
  expressions_filename = NA, start_date = lubridate::mdy("01/01/1800"),
  end_date = lubridate::mdy("12/31/2200"), month_nums = 1:12)
}
\arguments{
\item{model_name}{Filename of the iQuest neural network definition file.}

\item{scenario_name}{Optional label used to organize output (example: "A1B").}

\item{climate_source}{Optional label used to identify climate driver (example: "GFDL").}

\item{static_filename}{Filename containing the static site information.}

\item{dynamic_filename}{Filename containing the dynamic (climate) data.}

\item{dynamic_colnames}{Character vector containing column names of the dynamic datafile.}

\item{dynamic_filename2}{Optional second dynamic filename.}

\item{dynamic_colnames2}{Optional second dynamic filename column names.}

\item{field_order}{Optional numeric vector specifying an alternate run order.}

\item{expressions_filename}{Filename for the R expressions file, a file containing user-specified statistics
to calculate on the neural network model output.}

\item{start_date}{Earliest date of dynamic dataset that will be used in the simulation.}

\item{end_date}{Latest date of the dynamic dataset that will be used in the simulation.}

\item{month_nums}{Which month numbers to include in the simulations. c(6,7,8) would include June, July, and August only.}
}
\value{

}
\description{
This routine accepts the basics:
* neural network definition file (*.enn)
* start date
* end date
* climate source
* emissions scenario name
* static and dynamic data filenames
* static and dynamic column names
}
\details{
The result is a 'nnmodel object'--essentially an R list with all attributes needed to
use the functions contained in the rest of this package.
}
\examples{
model_filename   <- file.path(system.file("extdata/WI_StreamTemp.enn", package="nnlib") )
static_filename  <- file.path(system.file("extdata/static_data.csv", package="nnlib") )
dynamic_filename <- file.path(system.file("extdata/dynamic_data.csv", package="nnlib") )
mynnmodel <- nnmodel( model_name=model_filename,
                      scenario_name="Example problem",
                      climate_source="observed",
                      static_filename=static_filename,
                      dynamic_filename=dynamic_filename,
                      start_date=lubridate::mdy("01/01/1990"),
                      end_date=lubridate::mdy("12/31/2008") )
str( mynnmodel )

}

