% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnrun.R
\name{nnrun}
\alias{nnrun}
\title{Title}
\usage{
nnrun(nnmodel, static_site_col_name, static_site_name = "", mult_factor = 1,
  correction_factor_df = NA, correction_factor_col_name = NA,
  start_date = as.Date("01/01/1800", format = "\%m/\%d/\%Y"),
  end_date = as.Date("12/31/2200", format = "\%m/\%d/\%Y"))
}
\arguments{
\item{end_date}{}
}
\description{
Title
}

